REF: Refactor to remove the functional API in favor of DesktopBot class.
