--- docs/api/browser.md
@@ -1,5 +1,0 @@
-# Browser API
-
-## Functions
-
-::: botcity.core.browser
