--- docs/api/display.md
@@ -1,5 +1,0 @@
-# Display API
-
-## Functions
-
-::: botcity.core.display
