--- docs/api/index.md
@@ -1,23 +1,0 @@
-# API
-
-The API is divided into the following modules:
-
-- *display*: this module contains search/find functions that interact with the display 
-
-- *browser*: this module contains functions that interact with the web browser
-
-- *mouse*: this module contains functions that interact with the mouse
-
-- *keyboard*: this module contains functions that interact with the keyboard
-
-- *misc*: this module contains miscellaneous functions such as executing applications, and wait/print.
-
-- *bot*: this module contains the base class for Bots, including the DesktopBot
-
-In order to make it easy for users, all modules are available
-via the core module.
-```python
-from botcity.core import *
-```
-
-The following sections will describe each module in depth.
