--- docs/api/keyboard.md
@@ -1,5 +1,0 @@
-# Keyboard API
-
-## Functions
-
-::: botcity.core.keyboard
