--- docs/api/misc.md
@@ -1,5 +1,0 @@
-# Miscellaneous API
-
-## Functions
-
-::: botcity.core.misc
