--- docs/api/mouse.md
@@ -1,5 +1,0 @@
-# Mouse API
-
-## Functions
-
-::: botcity.core.mouse
