+++ mkdocs.yml
@@ -6,14 +6,8 @@ copyright: BotCity &copy; - All rights reserved
 nav:
     - Home: index.md
     - Getting Started: intro.md
-    - Framework:
-        - api/index.md
-        - 'Display': api/display.md
-        - 'Browser': api/browser.md
-        - 'Mouse': api/mouse.md
-        - 'Keyboard': api/keyboard.md
-        - 'Misc': api/misc.md
-        - 'Bot': api/bot.md
+    - Framework: bot.md
+
 theme:
   icon:
     repo: fontawesome/brands/github
